[DOC,FIX,TEST] address AdductInfo review thread
Agent-Logs-Url: https://github.com/OpenMS/OpenMS/sessions/c0ebf359-736c-4d65-bd47-2b3cac636508

+++ src/openms/source/CHEMISTRY/AdductInfo.cpp
@@ -189,7 +189,7 @@ namespace OpenMS
     // check if M has a multiplier in front
     if (m_part.length() > 1)
     { // will throw conversion error of not a number
-      mol_multiplier = m_part.prefix(m_part.length()-1).toDouble();
+      mol_multiplier = m_part.prefix(m_part.length() - 1).toInt();
     }
 
     // evaluate the adduct string ...
